Fix teleporting issue.
The issue is that, if a player teleported using like SetPlayerPos right, the next update time is set to +3 and the last position is never reset. So while the unix timestamp is equal to the next update time. Update the player's last position. Basically this will update the position a second before the check starts running again. Because the last position was never updated there pos was set manually.

Player is not checked while entering or exiting a vehicle.